requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers
pandas
joblib
scikit-learn
xgboost
Pillow
